refactor(xtensa): split esp32s3 glue into its own module
system/xtensa.rs (2128 lines) was a single file holding the generic Xtensa
CPU setup, the classic-ESP32 (LX6) builder, AND the ESP32-S3 builder. Split
the ESP32-S3 half (Esp32s3Opts/BootMode/Wiring/MemMap, configure_esp32s3_memmap,
configure_xtensa_esp32s3, register_esp32s3_peripherals, register_default_thunks,
ESP32S3_PERIPHERALS) into system/xtensa/esp32s3.rs.

xtensa.rs -> xtensa/mod.rs (1371 lines, generic + LX6 + shared RamPeripheral)
re-exporting esp32s3 so external paths (system::xtensa::configure_xtensa_esp32s3
etc.) are unchanged. Pure code move — workspace lib (1331) + register_coverage
+ firmware_survival (26) all green.
